move some classes to internal package
