--- openapi-codegen-http/src/main/java/org/davidmoten/oa3/codegen/http/service/HttpEntityOutputStream.java
@@ -1,5 +1,0 @@
-package org.davidmoten.oa3.codegen.http.service;
-
-public class HttpEntityOutputStream {
-
-}
