+++ openapi-codegen-http/src/main/java/org/davidmoten/oa3/codegen/http/service/StandardOption.java
@@ -2,5 +2,5 @@ package org.davidmoten.oa3.codegen.http.service;
 
 public enum StandardOption implements Option {
 
-    DISALLOW_PATCH;
+    PATCH_USING_HEADER;
 }
