+++ openapi-codegen-http/src/main/java/org/davidmoten/oa3/codegen/http/service/internal/DefaultHttpConnection.java
@@ -1,4 +1,4 @@
-package org.davidmoten.oa3.codegen.http.service;
+package org.davidmoten.oa3.codegen.http.service.internal;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -7,6 +7,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Consumer;
+
+import org.davidmoten.oa3.codegen.http.service.HttpConnection;
+import org.davidmoten.oa3.codegen.http.service.Response;
 
 public final class DefaultHttpConnection implements HttpConnection {
