+++ openapi-codegen-http/src/main/java/org/davidmoten/oa3/codegen/http/service/internal/DefaultResponse.java
@@ -1,10 +1,12 @@
-package org.davidmoten.oa3.codegen.http.service;
+package org.davidmoten.oa3.codegen.http.service.internal;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.util.List;
 import java.util.Map;
+
+import org.davidmoten.oa3.codegen.http.service.Response;
 
 public final class DefaultResponse implements Response {
